requirements.txt
dwave_system==1.24.0
matplotlib==3.10.7
moocore==0.1.8
numpy==2.3.4
